+++ database/migrations/2025_11_28_104042_add_m_recipes_column.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('m_recipes', function (Blueprint $table) {
-            $table->integer('category_id')->nullable()->comment('カテゴリID')->after('users_id');
+            $table->bigInteger('category_id')->nullable()->comment('カテゴリID')->after('users_id');
         });
     }
